% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_tree.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate tree with metadata}
\usage{
annotate_tree(tree, metadata, layout = "circular",
  tree_color = "black", line_width = 0.1, color_variable = NULL,
  tippoint_size = 1, label_variable = NULL, label_size = 4,
  label_offset = 0.01, clade_label_node = NULL, clade_label = NULL,
  cladelabel_offset = NULL, align_cladelabel = FALSE, align = FALSE,
  ladderize = TRUE, node_labels = NULL, palette_type = "Paired",
  legend_position = "right", own_palette = NULL)
}
\arguments{
\item{tree}{A tree object to annotate}

\item{metadata}{The path to a tab-separated metadata file with information to annotate with. The first column in the file must be exact-matching IDs to the IDs used to create the tree. May also be a data frame.}

\item{layout}{One of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight' (see \code{\link[ggtree]{ggtree}})}

\item{tree_color}{The color of the tree}

\item{line_width}{The width of the lines in the tree}

\item{color_variable}{A string representing the column name in the metadata file to plot as nodes on the tree, where each unique value in the column will be represented with a color}

\item{tippoint_size}{The size of the colored nodes in the tree}

\item{label_variable}{A string representing the column name in the metadata file to plot as labels on the tree}

\item{label_size}{The font size of the labels}

\item{label_offset}{The distance between the labels and the tree}

\item{clade_label_node}{The node where the clade label is placed}

\item{clade_label}{The label that will be placed on the clade}

\item{cladelabel_offset}{The distance between the tree and the clade label}

\item{align_cladelabel}{Should the clade label be aligned?}

\item{align}{Should the labels align?}

\item{ladderize}{Should the tips be ladderized?}

\item{node_labels}{Should node labels be included?}

\item{palette_type}{Specify which colorBrewer palette to use. Choose from "Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2", "Set3" (see \code{\link[RColorBrewer]{brewer.pal}})}

\item{legend_position}{Specify legend position in plot}

\item{own_palette}{If a named vector of colors are specified here, they are used instead of the automatically generated colors from colorBrewer}
}
\description{
Add informative node colors and labels to existing tree. The function creates unique color palettes based on unique column values in supplied metadata file.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
