% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distApply.R
\name{dist.apply}
\alias{dist.apply}
\title{Fill Vector with Values for a Distance Matrix}
\usage{
dist.apply(n, FUN)
}
\arguments{
\item{n}{the number of objects to compare}

\item{FUN}{the function to apply to all index pairs \code{i}, \code{j} of
objects to compare}
}
\value{
a vector of length \code{n(n-1)/2} with the results of \code{FUN}
}
\description{
Create a vector that can be used as backing store for a distance
  matrix by applying a function \code{FUN} to all indices that arise when
  comparing all the \code{n} objects.
}
