% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distApplySamplesRanked.R
\name{dist.apply.samples.ranked}
\alias{dist.apply.samples.ranked}
\title{Create a Distance Vector based on Normalized Ranking of Distances}
\usage{
dist.apply.samples.ranked(X, FUN = distance.euclidean, sampler = identity,
  rank.all = rank.dist, rank.fromSingle = identity, aggregate = mean,
  FUN.VALUE = +Inf, cores = 1L, logging = FALSE)
}
\arguments{
\item{X}{the list or vector of elements to be compared}

\item{FUN}{the distance function, a function accepting two \emph{samples} and
returning one distance value}

\item{sampler}{the sampling function, returning a vector or list of samples
for an element of \code{X}}

\item{rank.all}{the ranking to be applied to all distances}

\item{rank.fromSingle}{the ranking to be applied to all the distances from
one specific sample to the other samples}

\item{aggregate}{the aggregation function which will join distances computed
with \code{FUN} of all combinations of the samples from two elements of
\code{X} into a single value}

\item{FUN.VALUE}{the value to be used for situations where an element of
\code{X} contains no samples}

\item{cores}{the number of cores to be used for parallel computation}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
a vector of values that can be used to produce a distance matrix
}
\description{
Create a distance matrix/vector where the actual distances from
  samples are replaced by their mean ranks. This should allow for
  scale-independent, robust distance matrices.
}
