% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{dist.merge}
\alias{dist.merge}
\title{Merge Rows and Columns of a Distance Matrix According to a Selection}
\usage{
dist.merge(distObj, selection, names, aggregate = mean, cores = 1L,
  FUN.VALUE = +Inf)
}
\arguments{
\item{distObj}{the original distance object}

\item{selection}{the list of vectors selecting the matrix elements}

\item{names}{the new names to be used}

\item{aggregate}{the aggregate function to combine multiple distances}

\item{cores}{the number of cores to be used for the computation}

\item{FUN.VALUE}{the distance to be used if there are no distances for a
certain combination}
}
\value{
a new distance object representing the merged distances
}
\description{
A list \code{selection} of vectors of indexes of rows/columns
 into a distance matrix is provided. A new distance matrix is created where
 only these selected elements remain and where the distances are
 \code{aggregates} of the original values.

 Assume that a list \code{list(c(1,2), c(3,4))} is provided. This will create
 a distance matrix containing only a single distances, namely the one between
 the first and second group. This distance is then the \code{aggregate} of
 the original distances between \code{(1,3)}, \code{(1,4)}, \code{(2,3)}, and
 \code{(2,4)}. The aggregate is the \code{mean} by default, but could be any
 statistic.
}
\seealso{
dist.apply.samples
}
