% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{dist.n.from.dm}
\alias{dist.n.from.dm}
\title{Get the Number of Objects that were used to Build a Distance Matrix}
\usage{
dist.n.from.dm(distObj)
}
\arguments{
\item{distObj}{the distance object, matrix, or vector}
}
\value{
the number of objects whose distance can be represented in this
}
\description{
Given a distance matrix or plain vector \code{distObj}, find the
  number of objects whose distances can be represented by \code{distObj}.
  This function first tries to extract of \code{Size} attribute from the object.
  If that fails, it will use \code{\link{dist.n}} to compute the size based on
  the object's length.
}
\examples{
m <- matrix(rnorm(n=40), nrow=8)
d <- dist(m)
dist.n.from.dm(d)
# 8
}
\seealso{
dist.n
}
