% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{dist.n}
\alias{dist.n}
\title{Compute the Number of Objects from the Number of Slots}
\usage{
dist.n(slots)
}
\arguments{
\item{slots}{the number of slots in a distance matrix}
}
\value{
the number of objects whose distances are stored in the matrix
}
\description{
Given the number of \code{slots} of a distance matric, get the
  number of objects whose distance is computed. This is the inverse function
  to \code{\link{dist.slots}}.
}
\examples{
dist.slots(10)
# 45
dist.n(45)
# 10
}
\seealso{
dist.slots
}
