% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distance.euclidean}
\alias{distance.euclidean}
\title{A Function Computing the Euclidean Distance of Two Vectors}
\usage{
distance.euclidean(a, b)
}
\arguments{
\item{a}{the first vector}

\item{b}{the second vector}
}
\value{
the Euclidean distance between \code{a} and \code{b}
}
\description{
Well, a function computing the Euclidean distance between two
  vectors \code{a} and \code{b}.
}
