\name{distancer-package}
\alias{distancer-package}
\alias{distancer}
\docType{package}
\title{
\packageTitle{distancer}
}
\description{
\packageDescription{distancer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{distancer}
\packageIndices{distancer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{distancer}

Maintainer: \packageMaintainer{distancer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
