% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distify.R
\name{distify}
\alias{distify}
\title{Prepare snp-dist table}
\usage{
distify(df, colname = "snp-dists 0.6.3")
}
\arguments{
\item{df}{The data frame holding the results from snp-dists}

\item{colname}{The name of the first column in the data frame (usually snp-dists 0.6.3 or some other version)}
}
\description{
Function that prepares the snp-dists table for distance calculations
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
