% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_tip_labels.R
\name{fix_tip_labels}
\alias{fix_tip_labels}
\title{Fix tip labels in trees}
\usage{
fix_tip_labels(tree, df, sub_string, tree_type = "treedata")
}
\arguments{
\item{tree}{A tree object}

\item{df}{A data frame with two columns: ref - matching values to the tip points in the tree. id - the new tip values}

\item{sub_string}{The string to remove from the tip labels before matching values}

\item{tree_type}{The tree object type, either treedata or phylo}
}
\description{
Correct tip labels based on matching values to a separate data frame
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
