% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tanglegram.R
\name{plot_tanglegram}
\alias{plot_tanglegram}
\title{Create tanglegram from two dendrograms}
\usage{
plot_tanglegram(tree1, tree2, plot_title = NULL)
}
\arguments{
\item{tree1}{First tree object}

\item{tree2}{Second tree object}

\item{plot_title}{The title of the plot}
}
\description{
Compare two dendrograms by creating a tanglegram between the two trees. Note that the same node labels must be present in both trees
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
