% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a tree from a phylo object}
\usage{
plot_tree(phylo)
}
\arguments{
\item{phylo}{The phylo object to plot}
}
\description{
Function that plots a simple tree from a phylo object, with a treescale and tip labels
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
