% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_node.R
\name{prep_node}
\alias{prep_node}
\title{Prepare list of tips below node}
\usage{
prep_node(df, node)
}
\arguments{
\item{df}{The tree, in tibble format}

\item{node}{Numeric, the node of interest}
}
\description{
Function that identifies the tips below a given node, and outputs a data frame with the sample id and node id
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
