% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_node_stats.R
\name{prep_node_stats}
\alias{prep_node_stats}
\title{SNP distance statistics for given node}
\usage{
prep_node_stats(tree, dists, nodes, sub_string)
}
\arguments{
\item{tree}{Phylo, The tree}

\item{dists}{Data frame, The SNP distance matrix (from snp-dists)}

\item{nodes}{Numeric, vector of nodes of interest}

\item{sub_string}{The string that will be removed from the SNP distance matrix isolate names to match the tree tip labels}
}
\description{
Function that identifies the tips below a given node, calculates SNP distance statistics for all isolates below the given node, and outputs a data frame that can be used to plot with ggtree
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
