% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_diversity.R
\name{simpson_diversity}
\alias{simpson_diversity}
\title{Calculate Simpson diversity}
\usage{
simpson_diversity(df, colname)
}
\arguments{
\item{df}{Data frame holding the data in tidy format. For ST diversity calculations, have one column with the ST of each sample in the rows}

\item{colname}{The name of the column holding the data you want calculated. Must be given as a string.}
}
\description{
Function that calculates the Simpson diversity of data in given column
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
