% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_tree.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate tree with metadata}
\usage{
annotate_tree(tree, metadata, layout = "circular",
  tree_color = "black", line_width = 0.1, color_variable = NULL,
  tippoint_size = 1, label_variable = NULL, label_size = 4,
  label_offset = 0.01, shape_variable = NULL, shape_values = c(15,
  16, 17, 18, 19), clade_label_node = NULL, clade_label = NULL,
  cladelabel_offset = NULL, align_cladelabel = FALSE, align = FALSE,
  line_type = "dotted", line_size = 0.5, ladderize = TRUE,
  midroot = FALSE, node_labels = FALSE, bootstrap_lab = TRUE,
  nodelab_geom = "label", bootlab_size = 3, bootstrap_var = NULL,
  nodepoint_size = 3, nodepoint_color = "viridis",
  viridis_option = "D", treescale = TRUE, treescale_x = NULL,
  treescale_y = NULL, treescale_linesize = 0.5,
  palette_type = "Paired", legend_position = "right",
  color_palette = NULL, shape_palette = NULL, tree_type = "phylo")
}
\arguments{
\item{tree}{A tree object to annotate}

\item{metadata}{The path to a tab-separated metadata file with information to annotate with. The first column in the file must be exact-matching IDs to the IDs used to create the tree. May also be a data frame.}

\item{layout}{One of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight' (see \code{\link[ggtree]{ggtree}})}

\item{tree_color}{The color of the tree}

\item{line_width}{The width of the lines in the tree}

\item{color_variable}{A string representing the column name in the metadata file to plot as nodes on the tree, where each unique value in the column will be represented with a color}

\item{tippoint_size}{The size of the colored nodes in the tree}

\item{label_variable}{A string representing the column name in the metadata file to plot as labels on the tree}

\item{label_size}{The font size of the labels}

\item{label_offset}{The distance between the labels and the tree}

\item{shape_variable}{A string representing the column in the metadata file to represent as shapes on the tips}

\item{shape_values}{The shape types}

\item{clade_label_node}{The node where the clade label is placed}

\item{clade_label}{The label that will be placed on the clade}

\item{cladelabel_offset}{The distance between the tree and the clade label}

\item{align_cladelabel}{Should the clade label be aligned?}

\item{align}{Should the labels align?}

\item{line_type}{If align = TRUE, what type of line should be rendered?}

\item{line_size}{If align = TRUE, the size of the lines}

\item{ladderize}{Should the tips be ladderized?}

\item{midroot}{Should the tree be midrooted?}

\item{node_labels}{Should node labels be included? This is the node number, not the bootstrap values!}

\item{bootstrap_lab}{Should the bootstrap values be included? Note: This is the values in the node.label data from the tree}

\item{nodelab_geom}{Should the bootstrap labels be plain text or labels?}

\item{bootlab_size}{The size of the bootstrap values}

\item{bootstrap_var}{The bootstrap values used to create colored nodes in the tree, only works for Treedata types}

\item{nodepoint_size}{The size of the colored bootstrap nodes}

\item{nodepoint_color}{Color scale for bootstrap values, either "viridis" or "base"}

\item{treescale}{Should a treescale be rendered?}

\item{treescale_x}{x position of treescale}

\item{treescale_y}{y position of treescale}

\item{treescale_linesize}{Size of the treescale line}

\item{palette_type}{Specify which colorBrewer palette to use. Choose from "Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2", "Set3" (see \code{\link[RColorBrewer]{brewer.pal}})}

\item{legend_position}{Specify legend position in plot}

\item{color_palette}{If a named vector of colors are specified here, they are used instead of the automatically generated colors from colorBrewer}

\item{shape_palette}{Use your own named vector as shape palette}

\item{tree_type}{What type of tree object to use, either "treedata" or "phylo"}
}
\description{
Add informative node colors and labels to existing tree. The function creates unique color palettes based on unique column values in supplied metadata file.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
