% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_tree.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate tree with metadata}
\usage{
annotate_tree(tree, metadata_path, file = FALSE, layout = "circular",
  tree_color = "black", line_width = 0.1, color_variable,
  tippoint_size = 1, label_variable, label_size = 4,
  label_offset = 0.01, palette_type = "Paired", own_palette = NULL)
}
\arguments{
\item{tree}{A tree object to annotate}

\item{metadata_path}{The path to a tab-separated metadata file with information to annotate with. The first column in the file must be exact-matching IDs to the IDs used to create the tree}

\item{file}{Logical, should an R object be used as metadata file instead of a system file?}

\item{layout}{One of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight' (see \code{\link[ggtree]{ggtree}})}

\item{tree_color}{The color of the tree}

\item{line_width}{The width of the lines in the tree}

\item{color_variable}{A string representing the column name in the metadata file to plot as nodes on the tree, where each unique value in the column will be represented with a color}

\item{tippoint_size}{The size of the colored nodes in the tree}

\item{label_variable}{A string representing the column name in the metadata file to plot as labels on the tree}

\item{label_size}{The font size of the labels}

\item{label_offset}{The distance between the labels and the tree}

\item{palette_type}{Specify which colorBrewer palette to use. Choose from "Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2", "Set3" (see \code{\link[RColorBrewer]{brewer.pal}})}

\item{own_palette}{If a named vector of colors are specified here, they are used instead of the automatically generated colors from colorBrewer}
}
\description{
Add informative node colors and labels to existing tree. The function creates unique color palettes based on unique column values in supplied metadata file.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
