% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distances from cgMLST alleles}
\usage{
calc_dist(path, metric = "gower")
}
\arguments{
\item{path}{Full path to the cgMLST.tsv file, directly from chewBBACA}

\item{metric}{Character string specifying the metric to be used, default "gower" (see \code{\link[cluster]{daisy}})}
}
\description{
Imports the cgMLST.tsv file from chewBBACA and calculates distances between samples. Outputs a distance matrix.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
