% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distApply.R
\name{dist.apply}
\alias{dist.apply}
\title{Fill Vector with Values for a Distance Matrix}
\usage{
dist.apply(X, FUN = distance.euclidean, cores = 1L, logging = FALSE)
}
\arguments{
\item{X}{objects to compare}

\item{FUN}{the function to apply to all pairs of elements \code{a}, \code{b}
from \code{X}}

\item{cores}{the number of cores to be used for parallel computation}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
a vector of length \code{n(n-1)/2} with the results of \code{FUN}
}
\description{
Create a vector that can be used as backing store for a distance
  matrix by applying a function \code{FUN} to all pairings that arise when
  comparing all the objects in \code{X}.
}
