% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_simulation_cpp}
\alias{do_simulation_cpp}
\title{The function that does a simulation.}
\usage{
do_simulation_cpp(max_genetic_distance, mutation_rate, n_generations, n_loci,
  population_size, rng_seed, sampling_interval)
}
\arguments{
\item{max_genetic_distance}{the maximum number of different loci that two individuals can still produce viable offspring with}

\item{mutation_rate}{DNA mutation rate}

\item{n_generations}{number of generations}

\item{n_loci}{number of loci}

\item{population_size}{population size}

\item{rng_seed}{random number generator seed}

\item{sampling_interval}{after how many generations is the population sampled for species abundances}
}
\value{
Nothing
}
\description{
The function that does a simulation.
}

